Sorry, guys. Just a few final thoughts I had on the test

+++ src/main/java/com/narciso/tedtalks/speakers/dto/SpeakerInfluenceDto.java
@@ -1,7 +1,6 @@
 package com.narciso.tedtalks.speakers.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
-import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
@@ -12,8 +11,8 @@ import java.math.BigDecimal;
 public class SpeakerInfluenceDto {
     private Long speakerId;
     private String name;
-    private long totalViews;
-    private long totalLikes;
+    private int totalViews;
+    private int totalLikes;
     @JsonFormat(shape= JsonFormat.Shape.NUMBER_FLOAT, pattern="0.00")
     private BigDecimal score;
 }
